Use leaflet library to determine center point of the raster -Refactor utils funtions -Add types check for leaflet
