+++ package.json
@@ -8,6 +8,7 @@
     "@babel/core": "7.4.3",
     "@svgr/webpack": "4.1.0",
     "@types/jest": "^24.0.13",
+    "@types/leaflet": "^1.4.4",
     "@types/node": "^12.0.2",
     "@types/react": "^16.8.18",
     "@types/react-dom": "^16.8.4",
